Fix unicity checking of meta objects
