+++ doc/api.rst
@@ -18,11 +18,13 @@ Package API
     xoa.core.geo
     xoa.core.interp
     xoa.core.num
+    xoa.core.regrid
     xoa.data_samples
     xoa.dyn
     xoa.filter
     xoa.geo
     xoa.grid
+    xoa.interp
     xoa.krig
     xoa.meta
     xoa.misc
